% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dnelastic.R
\name{Dnelastic}
\alias{Dnelastic}
\title{Title}
\usage{
Dnelastic(X, eps, core_n = NULL, outfilename, verbose = FALSE)
}
\arguments{
\item{X}{???? row sample ???? matrix of features X}

\item{eps}{error}

\item{core_n}{Number of cpus for parallel computing, The default is 1}

\item{outfilename}{Output file location and name prefix}

\item{verbose}{Whether to output running information}
}
\description{
Title
}
\examples{

}
