% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dynetwork.R
\name{Dynetwork}
\alias{Dynetwork}
\title{Title}
\usage{
Dynetwork(X, connect, times, order, outfilename, verbose)
}
\arguments{
\item{X}{???? row sample ???? matrix of features X}

\item{connect}{error}

\item{times}{Number of cpus for parallel computing, The default is 1}

\item{order}{Output file location and name prefix}

\item{outfilename}{Output file location and name prefix}

\item{verbose}{Whether to output running information}
}
\description{
Title
}
\examples{

}
