% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynetwork_plot.R
\name{dynetwork_plot}
\alias{dynetwork_plot}
\title{Title}
\usage{
dynetwork_plot(Time, ori_eff_mat, times, connect, net_odee)
}
\arguments{
\item{Time}{???? row sample ???? matrix of features X}

\item{ori_eff_mat}{error}

\item{times}{Number of cpus for parallel computing, The default is 1}

\item{connect}{Whether to output running information}

\item{net_odee}{Whether to output running information}
}
\description{
Title
}
\examples{

}
