% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fclusty.R
\name{Fclusty}
\alias{Fclusty}
\title{Title}
\usage{
Fclusty(tmp_data, nt, J = 9, r = 0, init_rho, init_sigsq, out_file_name)
}
\arguments{
\item{tmp_data}{a Input data matrix}

\item{nt}{a Input data matrix}

\item{J}{a Input data matrix}

\item{r}{a Input data matrix}

\item{init_rho}{a Input data matrix}

\item{init_sigsq}{a Input data matrix}

\item{out_file_name}{a Input data matrix}
}
\description{
Title
}
