% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicNet_plot.R
\name{fclusty_plot}
\alias{fclusty_plot}
\title{Title}
\usage{
fclusty_plot(
  belong_data,
  ori_data,
  means_data,
  times,
  J,
  plot.control = list(y_cex = NULL, col_mean = "blue", main_label = NULL, no_out_pdf =
    TRUE)
)
}
\arguments{
\item{belong_data}{???? row sample ???? matrix of features X}

\item{ori_data}{error}

\item{means_data}{Number of cpus for parallel computing, The default is 1}

\item{times}{Output file location and name prefix}

\item{J}{Output file location and name prefix}

\item{plot.control}{Whether to output running information}
}
\description{
Title
}
\examples{

}
