% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DynamicNet_plot.R
\name{net_eff_plot}
\alias{net_eff_plot}
\title{Title}
\usage{
net_eff_plot(
  ori_eff_data,
  odee_matrix,
  times,
  plot.control = list(main_label = NULL)
)
}
\arguments{
\item{ori_eff_data}{???? row sample ???? matrix of features X}

\item{odee_matrix}{error}

\item{times}{Number of cpus for parallel computing, The default is 1}

\item{plot.control}{Whether to output running information}
}
\description{
Title
}
\examples{

}
