% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fclusty.R
\name{Fclusty}
\alias{Fclusty}
\title{Title}
\usage{
Fclusty(data, times, startK, endK, r, outfilename, verbose = FALSE)
}
\arguments{
\item{data}{Enter the observation data matrix with dimension ????, and each row is an observation vector.}

\item{times}{Enter the time series vector with the dimension ????}

\item{startK}{The initial number of clusters, of type integer, startK<endk< span=""></endk<>}

\item{endK}{Maximum number of clusters. Type is integer}

\item{r}{The order of an orthogonal polynomial of type integer}

\item{outfilename}{String data describing the location and name prefix of the output file, The general format is: \ output file path \ Output file name prefix}

\item{verbose}{Bool indicates whether to output running information}
}
\description{
Title
}
\examples{

}
